requirements.txt
fastapi
uvicorn
requests
apscheduler
aiohttp
numpy
python-dotenv
